Jenkinsfix

+++ src/core/properties/transferfunctionproperty.cpp
@@ -10,7 +10,7 @@ void TransferFunctionProperty::serialize(IvwSerializer& s) const {
 	Property::serialize(s) ;
 	TransferFunction temp = get();
 	std::stringstream stream;
-	s.serialize("size", temp.getNumberOfDataPoints());
+	s.serialize("size", (int)temp.getNumberOfDataPoints());
 
 	for (size_t i = 0; i < temp.getNumberOfDataPoints(); i++){
 		stream << "pos" << i;
